scheduler-backend/requirements.txt
# Python version requirement
# Python == 3.11.* (ortools requires Python 3.11, newer versions not supported)

# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-dateutil>=2.8.2
ortools>=9.3.10497

# Testing dependencies
pytest>=7.0.0
pytest-cov>=3.0.0  # For coverage reporting
pytest-xdist>=2.5.0  # For parallel test execution
psutil>=5.8.0  # For performance monitoring
freezegun>=1.2.0  # For date/time mocking in tests

# Development dependencies
black>=22.3.0  # Code formatting
isort>=5.10.1  # Import sorting
flake8>=4.0.1  # Linting
mypy>=0.950  # Type checking
